requirements.txt
event-scheduler==0.1.2
pkg_resources==0.0.0
PyYAML==5.4.1
